flask-ai/requirements.txt
Flask
flask-cors
requests
python-dotenv
fpdf
pymongo
PyGithub
llama-index
chromadb
google-generativeai
huggingface-hub
